# swagger.yaml
openapi: 3.0.0
info:
  title: swagger docs project
  description: a project for learning how to write swagger docs.
  version: 1.0.0
  contact:
    email: jaybhogayata53@gmail.com
    url: "https://www.codewithjay.co.in"

servers:
  - url: "{protocol}://localhost:8080/api/{version}"
    description: for localhost
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
      version:
        enum:
          - v1
          - v2
        default: v1

  - url: "{protocol}://swagger.codewithjay.co.in/api/{version}"
    description: for production
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
      version:
        enum:
          - v1
          - v2
        default: v1

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    BearerAuth:
      type: http
      scheme: bearer

paths:
  /:
    get:
      tags:
        - string
      summary: route for test the backend is up or not.
      responses:
        "200":
          description: A hello from backend.
          # what type of content in response
          content:
            application/json:
              schema:
                type: string
                example: "Hello"
        "400":
          description: bad request.
        "500":
          description: internal server error.

  /courseObject:
    get:
      tags:
        - object
      summary: return a course object.
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number

        "400":
          description: bad request.
        "500":
          description: internal server error.

  /courses:
    get:
      tags:
        - array
      summary: return a all course
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    price:
                      type: number

        "400":
          description: bad request.
        "500":
          description: internal server error.

  /course/{courseId}:
    get:
      tags:
        - params in request.
      summary: return a one course based on id.
      parameters:
        - name: courseId
          in: path
          required: true
          default: 11
          schema:
            type: string
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number
        "400":
          description: bad request.
        "500":
          description: internal server error.

  /addCourse:
    post:
      tags:
        - request body
      summary: add new course in a list.
      consumes:
        - application/json
      produces:
        - application/json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                price:
                  type: number
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                type: boolean
        "400":
          description: bad request.
        "500":
          description: internal server error.

  /coursequery:
    get:
      tags:
        - url query
      summary: how query work with swagger docs.
      parameters:
        - name: location
          in: query
          required: true
          schema:
            type: string
            enum: [nyc, del, blr]
        - name: device
          in: query
          required: true
          schema:
            type: string
            enum: [web, mobile]
      responses:
        "200":
          description: Ok
          contact:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    type: string
                  device:
                    type: string
        400:
          description: Bad request
        500:
          description: internal server error
  /courseupload:
    post:
      tags:
        - upload img
      summary: uploading course image
      parameters:
        - in: header
          name: auth
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: All good success

        400:
          description: Bad request
        500:
          description: internal server error